Implemented: When a User is deleted, the Notices registered by the User are deleted.

+++ src/Controller/UsersController.php
@@ -7,13 +7,16 @@ use Cake\Datasource\Exception\RecordNotFoundException;
 use Cake\Event\Event;
 use Cake\Http\Exception\NotFoundException;
 use Cake\I18n\FrozenTime;
+use Cake\ORM\TableRegistry;
+
 /**
  * Users Controller
  *
  * @property \App\Model\Table\UsersTable $Users
- *
+ * @property \App\Model\Table\NoticeTable $Notice;
  * @method \App\Model\Entity\User[]|\Cake\Datasource\ResultSetInterface paginate($object = null, array $settings = [])
  */
+
 class UsersController extends AppController
 {
     /**
@@ -172,13 +175,18 @@ class UsersController extends AppController
     {
         if($this->Auth->user('Type')=='マネージャー') {
             $this->request->allowMethod(['post', 'delete']);
+            $Notice= TableRegistry::get('Notice');
             $user = $this->Users->get($id);
-            if ($this->Users->delete($user)) {
-                $this->Flash->success(__('The user has been deleted.'));
+            $notice = $Notice->find()->where(['Author'=>$id])->all();
+            if($Notice->deleteMany($notice)) {
+                if ($this->Users->delete($user)) {
+                    $this->Flash->success(__('The user has been deleted.'));
+                } else {
+                    $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+                }
             } else {
                 $this->Flash->error(__('The user could not be deleted. Please, try again.'));
             }
-
             return $this->redirect(['action' => 'index']);
         } else {
             $this->Flash->error('You are not authorized.');
